feat #205: Nova estrutura do módulo de Perfil (`Profile`)
