+++ lib/src/app/modules/configuration/account/presenter/bloc/deactivate_account_bloc.dart
@@ -6,10 +6,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:is_it_safe_app/generated/l10n.dart';
 import 'package:is_it_safe_app/src/app/modules/auth/modules/login/presenter/pages/login_page.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/usecases/deactivate_account_use_case.dart';
 import 'package:is_it_safe_app/src/components/widgets/safe_snack_bar.dart';
 import 'package:is_it_safe_app/src/core/state/safe_stream.dart' as ss;
 import 'package:is_it_safe_app/src/core/interfaces/safe_bloc.dart';
-import 'package:is_it_safe_app/src/domain/use_case/deactivate_account_use_case.dart';
 
 class DeactivateAccountBloc extends SafeBloC {
   final deactivateAccountStream = ss.SafeStream<bool?>(data: null);
