+++ lib/src/app/modules/location/review/presenter/bloc/review_bloc.dart
@@ -2,13 +2,13 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:is_it_safe_app/generated/l10n.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/usecases/save_review_use_case.dart';
 import 'package:is_it_safe_app/src/components/config/safe_event.dart';
 import 'package:is_it_safe_app/src/core/constants/assets_constants.dart';
 import 'package:is_it_safe_app/src/core/constants/double_constants.dart';
 import 'package:is_it_safe_app/src/core/constants/string_constants.dart';
 import 'package:is_it_safe_app/src/core/interfaces/safe_bloc.dart';
 import 'package:is_it_safe_app/src/domain/entity/review_entity.dart';
-import 'package:is_it_safe_app/src/domain/use_case/save_review_use_case.dart';
 import 'package:result_dart/result_dart.dart';
 
 class ReviewBloc extends SafeBloC {
