+++ lib/src/app/modules/location/review/review_module.dart
@@ -3,8 +3,8 @@ import 'package:is_it_safe_app/src/app/modules/auth/services/auth_service.dart';
 import 'package:is_it_safe_app/src/app/modules/location/review/presenter/bloc/review_bloc.dart';
 import 'package:is_it_safe_app/src/app/modules/location/review/presenter/pages/review_page.dart';
 import 'package:is_it_safe_app/src/app/modules/location/services/location_service.dart';
-import 'package:is_it_safe_app/src/domain/use_case/save_review_use_case.dart';
-import 'package:is_it_safe_app/src/service/api/modules/profile/profile_service.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/usecases/save_review_use_case.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/services/profile_service.dart';
 
 class ReviewModule extends Module {
   @override
