+++ lib/src/app/modules/profile/domain/usecases/save_review_use_case.dart
@@ -1,8 +1,8 @@
 import 'package:is_it_safe_app/src/app/modules/location/error/safe_review_error.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/models/request/request_save_review.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/services/profile_service_interface.dart';
 import 'package:is_it_safe_app/src/core/interfaces/safe_use_case.dart';
 import 'package:is_it_safe_app/src/domain/entity/review_entity.dart';
-import 'package:is_it_safe_app/src/service/api/modules/profile/profile_service_interface.dart';
-import 'package:is_it_safe_app/src/service/api/modules/profile/request/request_save_review.dart';
 import 'package:result_dart/result_dart.dart';
 
 class SaveReviewUseCase extends SafeUseCase {
