+++ lib/src/app/modules/profile/domain/usecases/update_user_use_case.dart
@@ -1,10 +1,9 @@
+import 'package:is_it_safe_app/src/app/modules/configuration/account/error/safe_account_error.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/models/request/resquest_update_user.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/services/profile_service_interface.dart';
 import 'package:is_it_safe_app/src/core/interfaces/safe_use_case.dart';
-import 'package:is_it_safe_app/src/domain/entity/user_entity.dart';
-import 'package:is_it_safe_app/src/service/api/modules/profile/profile_service_interface.dart';
-import 'package:is_it_safe_app/src/service/api/modules/profile/request/resquest_update_user.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/models/user_entity.dart';
 import 'package:result_dart/result_dart.dart';
-
-import '../../app/modules/configuration/account/error/safe_account_error.dart';
 
 class UpdateUserUseCase extends SafeUseCase {
   final IProfileService _service;
