+++ lib/src/app/modules/profile/presenter/widgets/profile_header.dart
@@ -3,7 +3,7 @@ import 'package:is_it_safe_app/src/core/state/safe_builder.dart';
 import 'package:is_it_safe_app/src/core/state/safe_stream.dart';
 
 import '../../../../../components/widgets/safe_profile_header.dart';
-import '../../../../../domain/entity/user_entity.dart';
+import '../../domain/models/user_entity.dart';
 
 class ProfileHeader extends StatelessWidget {
   final SafeStream<UserEntity> user;
