+++ lib/src/app/modules/profile/profile_module.dart
@@ -1,12 +1,12 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:is_it_safe_app/src/app/modules/auth/services/auth_service.dart';
 import 'package:is_it_safe_app/src/app/modules/location/services/location_service.dart';
-import 'package:is_it_safe_app/src/domain/use_case/delete_review_use_case.dart';
-import 'package:is_it_safe_app/src/domain/use_case/get_user_use_case.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/usecases/delete_review_use_case.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/usecases/get_user_use_case.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/services/profile_service.dart';
 import 'package:is_it_safe_app/src/app/modules/profile/presenter/bloc/profile_bloc.dart';
 import 'package:is_it_safe_app/src/app/modules/profile/presenter/pages/profile_page.dart';
 import 'package:is_it_safe_app/src/domain/use_case/save_user_login_use_case.dart';
-import 'package:is_it_safe_app/src/service/api/modules/profile/profile_service.dart';
 import 'package:is_it_safe_app/src/service/shared_preferences/shared_preferences_service.dart';
 import 'package:is_it_safe_app/src/service/shared_preferences/shared_preferences_service_interface.dart';
