+++ test/app/modules/configuration_module/account_module/mocks/profile_service_mock.dart
@@ -1,4 +1,4 @@
-import 'package:is_it_safe_app/src/service/api/modules/profile/profile_service.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/services/profile_service.dart';
 import 'package:mocktail/mocktail.dart';
 
 class ProfileServiceMock extends Mock implements ProfileService {}
