+++ test/app/modules/location_module/review_module/domain/save_review_use_case_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:is_it_safe_app/src/domain/use_case/save_review_use_case.dart';
-import 'package:is_it_safe_app/src/service/api/modules/profile/request/request_save_review.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/models/request/request_save_review.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/usecases/save_review_use_case.dart';
 import 'package:modular_test/modular_test.dart';
 
 import '../mocks/mock_review_service.dart';
