+++ test/app/modules/profile_module/domain/get_user_use_case_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/models/response/response_get_user.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/usecases/get_user_use_case.dart';
 import 'package:is_it_safe_app/src/app/modules/profile/error/safe_profile_error.dart';
-import 'package:is_it_safe_app/src/domain/entity/user_entity.dart';
-import 'package:is_it_safe_app/src/domain/use_case/get_user_use_case.dart';
-import 'package:is_it_safe_app/src/service/api/modules/profile/response/response_get_user.dart';
+import 'package:is_it_safe_app/src/app/modules/profile/domain/models/user_entity.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../mocks/profile_service_mock.dart';
